refactor: verseDisplay 별도 feature로 분리
