--- src/features/drilling/components/verseDisplay/index.test.tsx
@@ -1,7 +1,0 @@
-import { describe, test } from 'vitest';
-
-describe('CardDisplay Test', () => {
-  test('renders loader when verse data query initializes', () => {});
-  test('renders error message when verse data query fails', () => {});
-  test('renders card swiper when verse data exist', () => {});
-});
