+++ src/features/verseDisplay/components/verseCardSwiper/index.test.tsx
